% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{Geocode address}
\usage{
geocode(address, api = "amap", key = NULL, output = "json")
}
\arguments{
\item{address}{a string, address to geocode, more specific the better}

\item{api}{which API to use, currently supports "Amap" or "Baidu"}

\item{key}{API application key}

\item{output}{specify output type, JSON or XML}
}
\value{
Geocode of address (latitude, longitude)  \emph{# take note of the order}
}
\description{
Geocode a particular address.
}
\examples{
\dontrun{
geocode("北京市朝阳区朝阳首府", api = "amap", key = "Your_Api_Key_Here")

# set application key via environment variables
file.edit("~/.Renviron")
# restart R to take effect
amap_ak <- Sys.getenv("amap_ak")
geocode("北京市朝阳区朝阳首府", key = amap_ak, output = "xml")
}

}
