% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geoconv}
\alias{geoconv}
\title{Geocode conversion}
\usage{
geoconv(lat, lon, key = NULL)
}
\arguments{
\item{lat}{geocode latitude, string or numeric}

\item{lon}{geocode longitude, string or numeric}

\item{key}{Baidu map API application key}
}
\value{
Baidu standard geocode (latitude, longitude) \emph{# take note of the order}
}
\description{
Convert non-Baidu standard geocode to Baidu standard geocode
}
\examples{
\dontrun{
geoconv("114.21892734521", "29.575429778924", key = "Your_Api_Key_Here")
}

}
