% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{get_endpoint}
\alias{get_endpoint}
\title{Provide API endpoints}
\usage{
get_endpoint(api)
}
\arguments{
\item{api}{string, which api to use?}
}
\value{
string, endpoint of api
}
\description{
Provide API endpoints
}
